v2.1.1

Approved-by: Juraj Mažári <mazari@netgrif.com>
Approved-by: Milan Mladoniczky <mladoniczky@netgrif.com>

+++ src/main/groovy/com/netgrif/workflow/startup/MongoDbRunner.groovy
@@ -25,9 +25,14 @@ class MongoDbRunner extends AbstractOrderedCommandLineRunner {
     @Value('${spring.data.mongodb.port}')
     private String port
 
+    @Value('${spring.data.mongodb.drop}')
+    private boolean dropDatabase
+
     @Override
     void run(String... strings) throws Exception {
-        log.info("Dropping Mongo database ${host}:${port}/${name}")
-        mongoTemplate.getDb().dropDatabase()
+        if (dropDatabase) {
+            log.info("Dropping Mongo database ${host}:${port}/${name}")
+            mongoTemplate.getDb().dropDatabase()
+        }
     }
 }
